+++ src/Model/WatchlistConfigModel.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace HeimrichHannot\WatchlistBundle\Model;
+
+use Contao\Model;
+
+class WatchlistConfigModel extends Model
+{
+    protected static $strTable = 'tl_watchlist_config';
+}