+++ src/Resources/contao/languages/de/modules.php
@@ -1,0 +1,6 @@
+<?php
+
+/**
+ * Backend modules
+ */
+$GLOBALS['TL_LANG']['MOD']['watchlist_config'][0] = 'Merklistenkonfigurationen';