+++ src/Resources/contao/languages/en/modules.php
@@ -5,3 +5,8 @@
  */
 
 $GLOBALS['TL_LANG']['FMD']['watchlist'] = array('Watchlist', 'Add content elements to an watchlist and download them at all.');
+
+/**
+ * Backend modules
+ */
+$GLOBALS['TL_LANG']['MOD']['watchlist_config'][0] = 'Merklistenkonfigurationen';