+++ src/Resources/contao/languages/en/tl_watchlist_config.php
@@ -1,0 +1,35 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_watchlist_config'];
+
+/**
+ * Fields
+ */
+$lang['tstamp'][0] = 'Revision date';
+$lang['title'][0] = 'Title';
+$lang['title'][1] = 'Please enter a title.';
+$lang['useDownloadLink'] = ['Show download link', 'Create a redirect link to a page providing all elements of an watchlist as download.'];
+$lang['downloadLink']    = ['Redirect page for download link', 'Redirect page providing all elements of an watchlist as download.'];
+$lang['downloadLinkUseNotification'] = ['Send downloadlink as notification', 'Select this option to send the download link as notificaiton (typically email).'];
+$lang['downloadLinkNotification'] = ['Download link notification', 'Choose the notification that should be used to send the download link.'];
+$lang['downloadLinkUseConfirmationNotification'] = ['Send confirmation notification', 'Select this option to send an confirmation notification to the user before sending the download link.'];
+$lang['downloadLinkFormConfigModule'] = ['Form configuration', 'Select the form module containing the form configuration. ATTENTION: If no configuration is selected, the user data will be used for shipping the notification.'];
+
+/**
+ * Legends
+ */
+$lang['general_legend'] = 'General settings';
+
+/**
+ * Buttons
+ */
+$lang['new'][0] = 'New Merklistenkonfiguration';
+$lang['new'][1] = 'Merklistenkonfiguration create';
+$lang['edit'][0] = 'Edit Merklistenkonfiguration';
+$lang['edit'][1] = 'Edit Merklistenkonfiguration ID %s';
+$lang['copy'][0] = 'Duplicate Merklistenkonfiguration';
+$lang['copy'][1] = 'Duplicate Merklistenkonfiguration ID %s';
+$lang['delete'][0] = 'Delete Merklistenkonfiguration';
+$lang['delete'][1] = 'Delete Merklistenkonfiguration ID %s';
+$lang['show'][0] = 'Merklistenkonfiguration details';
+$lang['show'][1] = 'Show the details of Merklistenkonfiguration ID %s';